scripts/requirements.txt
SPARQLWrapper==2.0.0
pandas==2.2.3
requests==2.32.3
numpy==1.26.4
psycopg2-binary==2.9.9
tqdm==4.66.2    
